.history/backend/requirements_20250630154954.txt
blinker==1.9.0
click==8.2.1
fal-client==0.7.0
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.1
requests==2.31.0
Werkzeug==3.1.3
